+++ src/create-resolver.ts
@@ -1,10 +1,11 @@
 import {
     resolveWithPrefix,
-    PrefixOptions,
+    resolveWithPrefixSync,
     ResolveOptions,
+    Options,
 } from './resolve-with-prefix';
 
-function createResolver(options: PrefixOptions & ResolveOptions = {}) {
+function createResolver(options: Options = {}) {
     return (packageId: string, { dirname }: ResolveOptions = {}) => {
         return resolveWithPrefix(packageId, {
             ...options,
@@ -13,4 +14,13 @@ function createResolver(options: PrefixOptions & ResolveOptions = {}) {
     };
 }
 
-export { createResolver };
+function createResolverSync(options: Options = {}) {
+    return (packageId: string, { dirname }: ResolveOptions = {}) => {
+        return resolveWithPrefixSync(packageId, {
+            ...options,
+            dirname,
+        });
+    };
+}
+
+export { createResolver, createResolverSync };