+++ src/resolve-with-prefix.test.ts
@@ -1,198 +1,372 @@
 import path from 'path';
 import { platform } from 'os';
 
-import { resolveWithPrefix } from './resolve-with-prefix';
+import {
+    resolveWithPrefix,
+    resolveWithPrefixSync,
+} from './resolve-with-prefix';
 
 const cwd = process.cwd();
+const dir = path.resolve(__dirname, '__sandbox__/app1/');
+
+beforeEach(() => {
+    process.chdir(dir);
+});
 
 afterEach(() => {
     process.chdir(cwd);
 });
 
-test('resolves full preset without prefix set', () => {
-    const dir = path.resolve(__dirname, '__sandbox__/app1/');
-    process.chdir(dir);
+describe('resolves full preset without prefix set', () => {
+    const checkResult = (result: string) => {
+        const expected = path.resolve(
+            dir,
+            'node_modules/one-preset-test/index.js',
+        );
+        expect(result).toEqual(expected);
+    };
 
-    const resolved = resolveWithPrefix('one-preset-test');
+    const packageId = 'one-preset-test';
 
-    const expected = path.resolve(dir, 'node_modules/one-preset-test/index.js');
-    expect(resolved).toEqual(expected);
-});
+    test('async', async () => {
+        const resolved = await resolveWithPrefix(packageId);
 
-test('resolves full preset with prefix set', () => {
-    const dir = path.resolve(__dirname, '__sandbox__/app1/');
-    process.chdir(dir);
-
-    const resolved = resolveWithPrefix('one-preset-test', {
-        prefix: 'one-preset',
+        checkResult(resolved);
     });
 
-    const expected = path.resolve(dir, 'node_modules/one-preset-test/index.js');
-    expect(resolved).toEqual(expected);
+    test('sync', () => {
+        const resolved = resolveWithPrefixSync(packageId);
+
+        checkResult(resolved);
+    });
 });
 
-test('resolves leading preset with prefix', () => {
-    const dir = path.resolve(__dirname, '__sandbox__/app1/');
-    process.chdir(dir);
+describe('resolves full preset with prefix set', () => {
+    const checkResult = (result: string) => {
+        const expected = path.resolve(
+            dir,
+            'node_modules/one-preset-test/index.js',
+        );
+        expect(result).toEqual(expected);
+    };
 
-    const resolved = resolveWithPrefix('test', {
-        prefix: 'one-preset',
+    const packageId = 'one-preset-test';
+    const options = { prefix: 'one-preset' };
+
+    test('async', async () => {
+        const resolved = await resolveWithPrefix(packageId, options);
+
+        checkResult(resolved);
     });
 
-    const expected = path.resolve(dir, 'node_modules/one-preset-test/index.js');
-    expect(resolved).toEqual(expected);
+    test('sync', () => {
+        const resolved = resolveWithPrefixSync(packageId, options);
+
+        checkResult(resolved);
+    });
 });
 
-test('resolves other without preset', () => {
-    const dir = path.resolve(__dirname, '__sandbox__/app1/');
-    process.chdir(dir);
+describe('resolves leading preset with prefix', () => {
+    const checkResult = (result: string) => {
+        const expected = path.resolve(
+            dir,
+            'node_modules/one-preset-test/index.js',
+        );
 
-    const resolved = resolveWithPrefix('other');
+        expect(result).toEqual(expected);
+    };
 
-    const expected = path.resolve(dir, 'node_modules/other/index.js');
-    expect(resolved).toEqual(expected);
-});
+    const packageId = 'test';
+    const options = { prefix: 'one-preset' };
 
-test('resolves leading preset with prefix over matching module', () => {
-    const dir = path.resolve(__dirname, '__sandbox__/app1/');
-    process.chdir(dir);
+    test('async', async () => {
+        const resolved = await resolveWithPrefix(packageId, options);
 
-    const resolved = resolveWithPrefix('other', { prefix: 'one-preset' });
-
-    const expected = path.resolve(
-        dir,
-        'node_modules/one-preset-other/index.js',
-    );
-    expect(resolved).toEqual(expected);
-});
-
-test('strict:true, module: works', () => {
-    const dir = path.resolve(__dirname, '__sandbox__/app1/');
-    process.chdir(dir);
-
-    const resolved = resolveWithPrefix('module:explicit-name-test', {
-        prefix: 'one-preset',
-        strict: true,
+        checkResult(resolved);
     });
 
-    const expected = path.resolve(
-        dir,
-        'node_modules/explicit-name-test/index.js',
-    );
-    expect(resolved).toEqual(expected);
+    test('sync', () => {
+        const resolved = resolveWithPrefixSync(packageId, options);
+
+        checkResult(resolved);
+    });
 });
 
-test('default strict:true, require module:', () => {
-    const dir = path.resolve(__dirname, '__sandbox__/app1/');
-    process.chdir(dir);
+describe('resolves other without preset', () => {
+    const checkResult = (result: string) => {
+        const expected = path.resolve(dir, 'node_modules/other/index.js');
 
-    try {
-        expect.hasAssertions();
+        expect(result).toEqual(expected);
+    };
 
-        resolveWithPrefix('explicit-name-test', { prefix: 'one-preset' });
-    } catch (error) {
+    const packageId = 'other';
+
+    test('async', async () => {
+        const resolved = await resolveWithPrefix(packageId);
+
+        checkResult(resolved);
+    });
+
+    test('sync', () => {
+        const resolved = resolveWithPrefixSync(packageId);
+
+        checkResult(resolved);
+    });
+});
+
+describe('resolves leading preset with prefix over matching module', () => {
+    const checkResult = (result: string) => {
+        const expected = path.resolve(
+            dir,
+            'node_modules/one-preset-other/index.js',
+        );
+
+        expect(result).toEqual(expected);
+    };
+
+    const packageId = 'other';
+    const options = { prefix: 'one-preset' };
+
+    test('async', async () => {
+        const resolved = await resolveWithPrefix(packageId, options);
+
+        checkResult(resolved);
+    });
+
+    test('sync', () => {
+        const resolved = resolveWithPrefixSync(packageId, options);
+
+        checkResult(resolved);
+    });
+});
+
+describe('strict:true, module: works', () => {
+    const checkResult = (result: string) => {
+        const expected = path.resolve(
+            dir,
+            'node_modules/explicit-name-test/index.js',
+        );
+
+        expect(result).toEqual(expected);
+    };
+
+    const packageId = 'module:explicit-name-test';
+    const options = { prefix: 'one-preset', strict: true };
+
+    test('async', async () => {
+        const resolved = await resolveWithPrefix(packageId, options);
+
+        checkResult(resolved);
+    });
+
+    test('sync', () => {
+        const resolved = resolveWithPrefixSync(packageId, options);
+
+        checkResult(resolved);
+    });
+});
+
+describe('strict:true, do not require module: when prefix not set', () => {
+    const checkResult = (result: string) => {
+        const expected = path.resolve(
+            dir,
+            'node_modules/explicit-name-test/index.js',
+        );
+
+        expect(result).toEqual(expected);
+    };
+
+    const packageId = 'explicit-name-test';
+    const options = { strict: true };
+
+    test('async', async () => {
+        const resolved = await resolveWithPrefix(packageId, options);
+
+        checkResult(resolved);
+    });
+
+    test('sync', () => {
+        const resolved = resolveWithPrefixSync(packageId, options);
+
+        checkResult(resolved);
+    });
+});
+
+describe('default strict:true, require module:', () => {
+    const checkError = (error: any) => {
         expect(error.code).toEqual('MODULE_NOT_FOUND');
         expect(error.message).toMatchInlineSnapshot(`
             "Cannot find module 'one-preset-explicit-name-test' from '<PROJECT_ROOT>'
             - If you want to resolve \\"explicit-name-test\\", use \\"module:explicit-name-test\\""
         `);
-    }
-});
+    };
 
-test('strict:true, require module:', () => {
-    const dir = path.resolve(__dirname, '__sandbox__/app1/');
-    process.chdir(dir);
+    const packageId = 'explicit-name-test';
+    const options = { prefix: 'one-preset' };
 
-    try {
+    test('async', async () => {
         expect.hasAssertions();
 
-        resolveWithPrefix('explicit-name-test', {
-            prefix: 'one-preset',
-            strict: true,
-        });
-    } catch (error) {
+        try {
+            await resolveWithPrefix(packageId, options);
+        } catch (error) {
+            checkError(error);
+        }
+    });
+
+    test('sync', () => {
+        expect.hasAssertions();
+
+        try {
+            resolveWithPrefixSync(packageId, options);
+        } catch (error) {
+            checkError(error);
+        }
+    });
+});
+
+describe('strict:true, require module:', () => {
+    const checkError = (error: any) => {
         expect(error.code).toEqual('MODULE_NOT_FOUND');
         expect(error.message).toMatchInlineSnapshot(`
             "Cannot find module 'one-preset-explicit-name-test' from '<PROJECT_ROOT>'
             - If you want to resolve \\"explicit-name-test\\", use \\"module:explicit-name-test\\""
         `);
-    }
-});
+    };
 
-test('strict:true, allow other errors', () => {
-    const dir = path.resolve(__dirname, '__sandbox__/app1/');
-    process.chdir(dir);
+    const packageId = 'explicit-name-test';
+    const options = { prefix: 'one-preset', strict: true };
 
-    try {
+    test('async', async () => {
         expect.hasAssertions();
-        resolveWithPrefix('one-preset-scoped', {
-            prefix: 'one-preset',
-            org: '@example',
-            strict: true,
-        });
-    } catch (error) {
-        expect(error.code).toEqual('MODULE_NOT_FOUND');
-        expect(error.message).toMatchInlineSnapshot(`
-            "Cannot find module 'one-preset-scoped' from '<PROJECT_ROOT>'
-            - Did you mean \\"@example/one-preset-scoped\\"?"
-        `);
-    }
-});
 
-test('strict:true, do not require module: when prefix not set', () => {
-    const dir = path.resolve(__dirname, '__sandbox__/app1/');
-    process.chdir(dir);
-
-    const resolved = resolveWithPrefix('explicit-name-test', {
-        strict: true,
+        try {
+            await resolveWithPrefix(packageId, options);
+        } catch (error) {
+            checkError(error);
+        }
     });
 
-    const expected = path.resolve(
-        dir,
-        'node_modules/explicit-name-test/index.js',
-    );
-    expect(resolved).toEqual(expected);
+    test('sync', () => {
+        expect.hasAssertions();
+
+        try {
+            resolveWithPrefixSync(packageId, options);
+        } catch (error) {
+            checkError(error);
+        }
+    });
 });
 
-test('if module not found but it looks like @org missing, add error', () => {
-    const dir = path.resolve(__dirname, '__sandbox__/app1/');
-    process.chdir(dir);
-
-    try {
-        expect.hasAssertions();
-        resolveWithPrefix('one-preset-scoped', {
-            prefix: 'one-preset',
-            org: '@example',
-        });
-    } catch (error) {
+describe('strict:true, allow other errors', () => {
+    const checkError = (error: any) => {
         expect(error.code).toEqual('MODULE_NOT_FOUND');
         expect(error.message).toMatchInlineSnapshot(`
             "Cannot find module 'one-preset-scoped' from '<PROJECT_ROOT>'
             - Did you mean \\"@example/one-preset-scoped\\"?"
         `);
-    }
+    };
+
+    const packageId = 'one-preset-scoped';
+    const options = { prefix: 'one-preset', org: '@example', strict: true };
+
+    test('async', async () => {
+        expect.hasAssertions();
+
+        try {
+            await resolveWithPrefix(packageId, options);
+        } catch (error) {
+            checkError(error);
+        }
+    });
+
+    test('sync', () => {
+        expect.hasAssertions();
+
+        try {
+            resolveWithPrefixSync(packageId, options);
+        } catch (error) {
+            checkError(error);
+        }
+    });
 });
 
-test('works with NODE_PATH', () => {
+describe('if module not found but it looks like @org missing, add error', () => {
+    const checkError = (error: any) => {
+        expect(error.code).toEqual('MODULE_NOT_FOUND');
+        expect(error.message).toMatchInlineSnapshot(`
+            "Cannot find module 'one-preset-scoped' from '<PROJECT_ROOT>'
+            - Did you mean \\"@example/one-preset-scoped\\"?"
+        `);
+    };
+
+    const packageId = 'one-preset-scoped';
+    const options = { prefix: 'one-preset', org: '@example' };
+
+    test('async', async () => {
+        expect.hasAssertions();
+
+        try {
+            await resolveWithPrefix(packageId, options);
+        } catch (error) {
+            checkError(error);
+        }
+    });
+
+    test('sync', () => {
+        expect.hasAssertions();
+
+        try {
+            resolveWithPrefixSync(packageId, options);
+        } catch (error) {
+            checkError(error);
+        }
+    });
+});
+
+describe('works with NODE_PATH', () => {
     const nodePath = process.env.NODE_PATH;
-    const dir = path.resolve(__dirname, '__sandbox__/app1/node_modules');
+    const dirNodeModules = path.resolve(dir, 'node_modules');
 
-    const sep = platform() === 'win32' ? ';' : ':';
-    if (process.env.NODE_PATH) {
-        const splitNodePath = process.env.NODE_PATH.split(sep);
-        splitNodePath.push(dir);
+    beforeEach(() => {
+        process.chdir(cwd);
 
-        process.env.NODE_PATH = splitNodePath.join(':');
-    } else {
-        process.env.NODE_PATH = dir;
-    }
+        const sep = platform() === 'win32' ? ';' : ':';
+        if (process.env.NODE_PATH) {
+            const splitNodePath = process.env.NODE_PATH.split(sep);
+            splitNodePath.push(dirNodeModules);
 
-    const resolved = resolveWithPrefix('one-preset-test', {
-        prefix: 'one-preset',
+            process.env.NODE_PATH = splitNodePath.join(':');
+        } else {
+            process.env.NODE_PATH = dirNodeModules;
+        }
     });
 
-    const expected = path.resolve(dir, 'one-preset-test/index.js');
+    afterEach(() => {
+        process.env.NODE_PATH = nodePath;
+    });
 
-    expect(resolved).toEqual(expected);
-    process.env.NODE_PATH = nodePath;
+    const checkResult = (result: string) => {
+        const expected = path.resolve(
+            dirNodeModules,
+            'one-preset-test/index.js',
+        );
+
+        expect(result).toEqual(expected);
+    };
+
+    const packageId = 'one-preset-test';
+    const options = { prefix: 'one-preset' };
+
+    test('async', async () => {
+        const resolved = await resolveWithPrefix(packageId, options);
+
+        checkResult(resolved);
+    });
+
+    test('sync', () => {
+        const resolved = resolveWithPrefixSync(packageId, options);
+
+        checkResult(resolved);
+    });
 });